requirements.txt
asgiref==3.4.1
boto3==1.18.14
botocore==1.21.14
braintree==4.11.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
jmespath==0.10.0
Pillow==8.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
